Changing tests to work on Windows
