+++ source/Calamari.Tests/Fixtures/PowerShell/Scripts/PowerShellProgressPreferenceCheck.ps1
@@ -1,0 +1,1 @@
+﻿Write-Host "ProgressPreference = ${ProgressPreference}"
